requirements.txt
# FEPD - Forensic Evidence Parser Dashboard
# Unified Python Package Requirements
# Install: pip install -r requirements.txt

# ============================================================================
# GUI FRAMEWORK
# ============================================================================
PyQt6>=6.4.0,<7.0.0
PyQt6-WebEngine>=6.4.0
PyQt6-Charts>=6.4.0

# ============================================================================
# DATA ANALYSIS
# ============================================================================
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ============================================================================
# ADVANCED VISUALIZATIONS
# ============================================================================
plotly>=5.14.0
squarify>=0.4.3

# ============================================================================
# ARTIFACT PARSERS
# ============================================================================
python-evtx>=0.7.4
python-registry>=1.3.1
python-dateutil>=2.8.0

# ============================================================================
# CONFIGURATION
# ============================================================================
pyyaml>=6.0
python-dotenv>=1.0.0

# ============================================================================
# DISK IMAGE HANDLING & FORENSIC ANALYSIS
# ============================================================================
pytsk3>=20230125
libewf-python>=20230212

# ============================================================================
# PDF REPORT GENERATION
# ============================================================================
reportlab>=4.0.0
pillow>=10.0.0
qrcode>=7.4.2

# ============================================================================
# NETWORK VISUALIZATION
# ============================================================================
networkx>=3.1

# ============================================================================
# MACHINE LEARNING & TRAINING
# ============================================================================
scikit-learn>=1.3.0
tensorflow>=2.15.0
imbalanced-learn>=0.11.0
optuna>=3.4.0

# ============================================================================
# TEMPLATE ENGINE
# ============================================================================
jinja2>=3.1.0

# ============================================================================
# SYSTEM & MONITORING
# ============================================================================
psutil>=5.9.0

# ============================================================================
# FEPD TERMINAL (Interactive Forensic Shell)
# ============================================================================
prompt_toolkit>=3.0.36
pygments>=2.16.1

# ============================================================================
# RAG CHATBOT
# ============================================================================
sentence-transformers>=2.2.0
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
